requirements.txt
apache-airflow
pandas
requests
apache-airflow-providers-sqlite